Refactored code. Matched coding styles.
Put creation of required extension list into separate function. Also,
editted checkExtensionSupport function to use foreach loops to be the same
as the rest of the program. Also, made checkExtensionSupport to take in a
vector instead of an int and a pointer to an array to better match the
rest of the program.