Added validation layer callback.
Weird behavior:
    When the program is closing, it realizes the debug
callback extension object is not deleted and calls the callback function.
The issue is that this callback function is called 4 times on the same
object. This is not a huge issue and currently does not pose a threat in
the immediate future, but it may be worth looking into.