Added index buffer use.
We still currently allocate memory for each buffer which is not the right
way to do it. Even on recent GPUs you only get about 4000 simultaneous
allocations. Also, should try storing vertex and index buffers into a
single buffer for cache friendliness or even try to reuse memory chunks
for multiple resources if not used during the same render operations.