Add auto build batch file.
I tried to make a batch file to automatically build the vulkan shaders,
but there is an issue with the provided shader compiler. It takes the file
extension and makes that the name of the file, but the compiler requires
all of a shader type to have a corresponding file extension. This means
that you can't currently compile several vertex shaders and only the last
one compiled will be the outputed vert.spv.